requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
astroid==3.3.11
autopep8==2.3.2
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
Django==5.2.8
djangorestframework==3.16.1
flake8==7.3.0
fonttools==4.60.1
idna==3.11
isort==6.1.0
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.7
mccabe==0.7.0
mplsoccer==1.6.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycodestyle==2.14.0
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
